Move the admin contracts and their tests to the correct path
