+++ packages/airnode-protocol/contracts/authorizers/admin/AirnodeTokenLock.sol
@@ -2,7 +2,7 @@
 pragma solidity 0.8.9;
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import "../../../authorizers/interfaces/IRequesterAuthorizerWithManager.sol";
+import "../interfaces/IRequesterAuthorizerWithManager.sol";
 import "./AirnodeRequesterAuthorizerRegistryClient.sol";
 import "./AirnodeFeeRegistryClient.sol";
 import "./AirnodeTokenLockRolesWithManager.sol";
