+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeFeeRegistry.sol
