+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeFeeRegistryClient.sol
