+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeFeeRegistryRolesWithManager.sol
