+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeRequesterAuthorizerRegistry.sol
