+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeRequesterAuthorizerRegistryClient.sol
