+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeTokenLock.sol
