+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeTokenLockRolesWithManager.sol
