+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeTokenPayment.sol
