+++ packages/airnode-protocol/contracts/authorizers/admin/interfaces/IAirnodeTokenPaymentRolesWithManager.sol
