+++ packages/airnode-protocol/contracts/authorizers/admin/mock/MockApi3Token.sol
