+++ packages/airnode-protocol/test/authorizers/RequesterAuthorizerWithAirnode.sol.js
@@ -1,135 +1,125 @@
 /* globals context */
 const hre = require('hardhat');
 const { expect } = require('chai');
-const utils = require('../../utils');
+const utils = require('../utils');
 
 let roles;
-let accessControlRegistry, requesterAuthorizerWithManager;
-let requesterAuthorizerWithManagerAdminRoleDescription = 'RequesterAuthorizerWithManager admin';
+let accessControlRegistry, requesterAuthorizerWithAirnode;
+let requesterAuthorizerWithAirnodeAdminRoleDescription = 'RequesterAuthorizerWithAirnode admin';
 let adminRole, whitelistExpirationExtenderRole, whitelistExpirationSetterRole, indefiniteWhitelisterRole;
-let airnodeAddress = utils.generateRandomAddress();
+let airnodeAddress, airnodeMnemonic, airnodeWallet;
 let endpointId = utils.generateRandomBytes32();
 
 beforeEach(async () => {
   const accounts = await hre.ethers.getSigners();
   roles = {
     deployer: accounts[0],
-    manager: accounts[1],
-    whitelistExpirationExtender: accounts[2],
-    whitelistExpirationSetter: accounts[3],
-    indefiniteWhitelister: accounts[4],
-    anotherIndefiniteWhitelister: accounts[5],
+    whitelistExpirationExtender: accounts[1],
+    whitelistExpirationSetter: accounts[2],
+    indefiniteWhitelister: accounts[3],
+    anotherIndefiniteWhitelister: accounts[4],
     requester: accounts[5],
     randomPerson: accounts[9],
   };
   const accessControlRegistryFactory = await hre.ethers.getContractFactory('AccessControlRegistry', roles.deployer);
   accessControlRegistry = await accessControlRegistryFactory.deploy();
-  const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
-    'RequesterAuthorizerWithManager',
+  const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
+    'RequesterAuthorizerWithAirnode',
     roles.deployer
   );
-  requesterAuthorizerWithManager = await requesterAuthorizerWithManagerFactory.deploy(
+  requesterAuthorizerWithAirnode = await requesterAuthorizerWithAirnodeFactory.deploy(
     accessControlRegistry.address,
-    requesterAuthorizerWithManagerAdminRoleDescription,
-    roles.manager.address
+    requesterAuthorizerWithAirnodeAdminRoleDescription
   );
-  const managerRootRole = await accessControlRegistry.deriveRootRole(roles.manager.address);
-  adminRole = await requesterAuthorizerWithManager.adminRole();
-  whitelistExpirationExtenderRole = await requesterAuthorizerWithManager.whitelistExpirationExtenderRole();
-  whitelistExpirationSetterRole = await requesterAuthorizerWithManager.whitelistExpirationSetterRole();
-  indefiniteWhitelisterRole = await requesterAuthorizerWithManager.indefiniteWhitelisterRole();
-  await accessControlRegistry.connect(roles.manager).initializeAndGrantRoles(
-    [managerRootRole, adminRole, adminRole, adminRole, adminRole],
+  ({ airnodeAddress: airnodeAddress, airnodeMnemonic: airnodeMnemonic } = utils.generateRandomAirnodeWallet());
+  await roles.deployer.sendTransaction({
+    to: airnodeAddress,
+    value: hre.ethers.utils.parseEther('1'),
+  });
+  airnodeWallet = hre.ethers.Wallet.fromMnemonic(airnodeMnemonic).connect(hre.ethers.provider);
+  const airnodeRootRole = await accessControlRegistry.deriveRootRole(airnodeAddress);
+  adminRole = await requesterAuthorizerWithAirnode.deriveAdminRole(airnodeAddress);
+  whitelistExpirationExtenderRole = await requesterAuthorizerWithAirnode.deriveWhitelistExpirationExtenderRole(
+    airnodeAddress
+  );
+  whitelistExpirationSetterRole = await requesterAuthorizerWithAirnode.deriveWhitelistExpirationSetterRole(
+    airnodeAddress
+  );
+  indefiniteWhitelisterRole = await requesterAuthorizerWithAirnode.deriveIndefiniteWhitelisterRole(airnodeAddress);
+  await accessControlRegistry.connect(airnodeWallet).initializeAndGrantRoles(
+    [airnodeRootRole, adminRole, adminRole, adminRole, adminRole],
     [
-      requesterAuthorizerWithManagerAdminRoleDescription,
-      await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
-      await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
-      await requesterAuthorizerWithManager.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
-      await requesterAuthorizerWithManager.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
+      requesterAuthorizerWithAirnodeAdminRoleDescription,
+      await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
+      await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
+      await requesterAuthorizerWithAirnode.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
+      await requesterAuthorizerWithAirnode.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
     ],
     [
-      roles.manager.address, // which will already have been granted the role
+      airnodeAddress, // which will already have been granted the role
       roles.whitelistExpirationExtender.address,
       roles.whitelistExpirationSetter.address,
       roles.indefiniteWhitelister.address,
       roles.anotherIndefiniteWhitelister.address,
-    ]
+    ],
+    { gasLimit: 1000000 }
   );
   // Grant `roles.randomPerson` some invalid roles
   await accessControlRegistry
-    .connect(roles.manager)
+    .connect(airnodeWallet)
     .initializeAndGrantRoles(
-      [managerRootRole, managerRootRole, managerRootRole, managerRootRole],
+      [airnodeRootRole, airnodeRootRole, airnodeRootRole, airnodeRootRole],
       [
         Math.random(),
-        await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
-        await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
-        await requesterAuthorizerWithManager.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
+        await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
+        await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
+        await requesterAuthorizerWithAirnode.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
       ],
-      [roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address]
+      [roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address],
+      { gasLimit: 1000000 }
     );
 });
 
 describe('constructor', function () {
   context('AccessControlRegistry address is not zero', function () {
     context('Admin role description string is not empty', function () {
-      context('Manager address is not zero', function () {
-        it('constructs', async function () {
-          const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
-            'RequesterAuthorizerWithManager',
-            roles.deployer
-          );
-          requesterAuthorizerWithManager = await requesterAuthorizerWithManagerFactory.deploy(
-            accessControlRegistry.address,
-            requesterAuthorizerWithManagerAdminRoleDescription,
-            roles.manager.address
-          );
-          expect(await requesterAuthorizerWithManager.accessControlRegistry()).to.equal(accessControlRegistry.address);
-          expect(await requesterAuthorizerWithManager.adminRoleDescription()).to.equal(
-            requesterAuthorizerWithManagerAdminRoleDescription
-          );
-          expect(await requesterAuthorizerWithManager.manager()).to.equal(roles.manager.address);
-        });
-      });
-      context('Manager address is zero', function () {
-        it('reverts', async function () {
-          const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
-            'RequesterAuthorizerWithManager',
-            roles.deployer
-          );
-          await expect(
-            requesterAuthorizerWithManagerFactory.deploy(
-              accessControlRegistry.address,
-              requesterAuthorizerWithManagerAdminRoleDescription,
-              hre.ethers.constants.AddressZero
-            )
-          ).to.be.revertedWith('Manager address zero');
-        });
+      it('constructs', async function () {
+        const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
+          'RequesterAuthorizerWithAirnode',
+          roles.deployer
+        );
+        requesterAuthorizerWithAirnode = await requesterAuthorizerWithAirnodeFactory.deploy(
+          accessControlRegistry.address,
+          requesterAuthorizerWithAirnodeAdminRoleDescription
+        );
+        expect(await requesterAuthorizerWithAirnode.accessControlRegistry()).to.equal(accessControlRegistry.address);
+        expect(await requesterAuthorizerWithAirnode.adminRoleDescription()).to.equal(
+          requesterAuthorizerWithAirnodeAdminRoleDescription
+        );
       });
     });
     context('Admin role description string is empty', function () {
       it('reverts', async function () {
-        const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
-          'RequesterAuthorizerWithManager',
+        const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
+          'RequesterAuthorizerWithAirnode',
           roles.deployer
         );
         await expect(
-          requesterAuthorizerWithManagerFactory.deploy(accessControlRegistry.address, '', roles.manager.address)
+          requesterAuthorizerWithAirnodeFactory.deploy(accessControlRegistry.address, '')
         ).to.be.revertedWith('Admin role description empty');
       });
     });
   });
   context('AccessControlRegistry address is zero', function () {
     it('reverts', async function () {
-      const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
-        'RequesterAuthorizerWithManager',
+      const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
+        'RequesterAuthorizerWithAirnode',
         roles.deployer
       );
       await expect(
-        requesterAuthorizerWithManagerFactory.deploy(
+        requesterAuthorizerWithAirnodeFactory.deploy(
           hre.ethers.constants.AddressZero,
-          requesterAuthorizerWithManagerAdminRoleDescription,
-          roles.manager.address
+          requesterAuthorizerWithAirnodeAdminRoleDescription
         )
       ).to.be.revertedWith('ACR address zero');
     });
@@ -141,7 +131,7 @@ describe('extendWhitelistExpiration', function () {
     context('Timestamp extends whitelist expiration', function () {
       it('extends whitelist expiration', async function () {
         let whitelistStatus;
-        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -150,11 +140,11 @@ describe('extendWhitelistExpiration', function () {
         expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
         const expirationTimestamp = 1000;
         await expect(
-          requesterAuthorizerWithManager
+          requesterAuthorizerWithAirnode
             .connect(roles.whitelistExpirationExtender)
             .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
         )
-          .to.emit(requesterAuthorizerWithManager, 'ExtendedWhitelistExpiration')
+          .to.emit(requesterAuthorizerWithAirnode, 'ExtendedWhitelistExpiration')
           .withArgs(
             airnodeAddress,
             endpointId,
@@ -162,7 +152,7 @@ describe('extendWhitelistExpiration', function () {
             roles.whitelistExpirationExtender.address,
             expirationTimestamp
           );
-        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -174,21 +164,21 @@ describe('extendWhitelistExpiration', function () {
     context('Timestamp does not extend whitelist expiration', function () {
       it('reverts', async function () {
         await expect(
-          requesterAuthorizerWithManager
+          requesterAuthorizerWithAirnode
             .connect(roles.whitelistExpirationExtender)
             .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
         ).to.be.revertedWith('Does not extend expiration');
       });
     });
   });
-  context('Sender is the manager address', function () {
+  context('Sender is the Airnode address', function () {
     context('Timestamp extends whitelist expiration', function () {
       it('extends whitelist expiration', async function () {
         await accessControlRegistry
-          .connect(roles.manager)
-          .renounceRole(whitelistExpirationExtenderRole, roles.manager.address);
+          .connect(airnodeWallet)
+          .renounceRole(whitelistExpirationExtenderRole, airnodeAddress, { gasLimit: 1000000 });
         let whitelistStatus;
-        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -197,13 +187,15 @@ describe('extendWhitelistExpiration', function () {
         expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
         const expirationTimestamp = 1000;
         await expect(
-          requesterAuthorizerWithManager
-            .connect(roles.manager)
-            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
+          requesterAuthorizerWithAirnode
+            .connect(airnodeWallet)
+            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp, {
+              gasLimit: 1000000,
+            })
         )
-          .to.emit(requesterAuthorizerWithManager, 'ExtendedWhitelistExpiration')
-          .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, expirationTimestamp);
-        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+          .to.emit(requesterAuthorizerWithAirnode, 'ExtendedWhitelistExpiration')
+          .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, expirationTimestamp);
+        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -215,30 +207,30 @@ describe('extendWhitelistExpiration', function () {
     context('Timestamp does not extend whitelist expiration', function () {
       it('reverts', async function () {
         await accessControlRegistry
-          .connect(roles.manager)
-          .renounceRole(whitelistExpirationExtenderRole, roles.manager.address);
+          .connect(airnodeWallet)
+          .renounceRole(whitelistExpirationExtenderRole, airnodeAddress, { gasLimit: 1000000 });
         await expect(
-          requesterAuthorizerWithManager
-            .connect(roles.manager)
-            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
+          requesterAuthorizerWithAirnode
+            .connect(airnodeWallet)
+            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0, { gasLimit: 1000000 })
         ).to.be.revertedWith('Does not extend expiration');
       });
     });
   });
-  context('Sender does not have the whitelist extender role and is not the manager address', function () {
+  context('Sender does not have the whitelist extender role and is not the Airnode address', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 1000)
       ).to.be.revertedWith('Not expiration extender');
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 1000)
       ).to.be.revertedWith('Not expiration extender');
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.randomPerson)
           .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 1000)
       ).to.be.revertedWith('Not expiration extender');
@@ -252,11 +244,11 @@ describe('setWhitelistExpiration', function () {
       let whitelistStatus;
       const expirationTimestamp = 1000;
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
+        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
         .withArgs(
           airnodeAddress,
           endpointId,
@@ -264,7 +256,7 @@ describe('setWhitelistExpiration', function () {
           roles.whitelistExpirationSetter.address,
           expirationTimestamp
         );
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -272,13 +264,13 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(expirationTimestamp);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
+        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.whitelistExpirationSetter.address, 0);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -287,21 +279,23 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
     });
   });
-  context('Sender is tha manager address', function () {
+  context('Sender is the Airnode address', function () {
     it('sets whitelist expiration', async function () {
       await accessControlRegistry
-        .connect(roles.manager)
-        .renounceRole(whitelistExpirationSetterRole, roles.manager.address);
+        .connect(airnodeWallet)
+        .renounceRole(whitelistExpirationSetterRole, airnodeAddress, { gasLimit: 1000000 });
       let whitelistStatus;
       const expirationTimestamp = 1000;
       await expect(
-        requesterAuthorizerWithManager
-          .connect(roles.manager)
-          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
+        requesterAuthorizerWithAirnode
+          .connect(airnodeWallet)
+          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp, {
+            gasLimit: 1000000,
+          })
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, expirationTimestamp);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, expirationTimestamp);
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -309,13 +303,13 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(expirationTimestamp);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       await expect(
-        requesterAuthorizerWithManager
-          .connect(roles.manager)
-          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
+        requesterAuthorizerWithAirnode
+          .connect(airnodeWallet)
+          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0, { gasLimit: 1000000 })
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, 0);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, 0);
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -324,20 +318,20 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
     });
   });
-  context('Sender does not have the whitelist expiration setter role and is not the manager address', function () {
+  context('Sender does not have the whitelist expiration setter role and is not the Airnode address', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationExtender)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       ).to.be.revertedWith('Not expiration setter');
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       ).to.be.revertedWith('Not expiration setter');
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.randomPerson)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       ).to.be.revertedWith('Not expiration setter');
@@ -351,13 +345,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       let whitelistStatus;
       // Whitelist indefinitely
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, true, 1);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -365,7 +359,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -374,13 +368,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(true);
       // Whitelisting indefinitely again should have no effect
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, true, 1);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -388,7 +382,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -397,13 +391,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(true);
       // Revoke indefinite whitelisting
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, false, 0);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -411,7 +405,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -420,13 +414,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(false);
       // Revoking indefinite whitelisting again should have no effect
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, false, 0);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -434,7 +428,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -443,19 +437,23 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(false);
     });
   });
-  context('Sender is the manager address', function () {
+  context('Sender is the Airnode address', function () {
     it('sets indefinite whitelist status', async function () {
-      await accessControlRegistry.connect(roles.manager).renounceRole(indefiniteWhitelisterRole, roles.manager.address);
+      await accessControlRegistry
+        .connect(airnodeWallet)
+        .renounceRole(indefiniteWhitelisterRole, airnodeAddress, { gasLimit: 1000000 });
       let whitelistStatus;
       // Whitelist indefinitely
       await expect(
-        requesterAuthorizerWithManager
-          .connect(roles.manager)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
+        requesterAuthorizerWithAirnode
+          .connect(airnodeWallet)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true, {
+            gasLimit: 1000000,
+          })
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, true, 1);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, true, 1);
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -463,22 +461,24 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          roles.manager.address
+          airnodeAddress
         )
       ).to.equal(true);
       // Whitelisting indefinitely again should have no effect
       await expect(
-        requesterAuthorizerWithManager
-          .connect(roles.manager)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
+        requesterAuthorizerWithAirnode
+          .connect(airnodeWallet)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true, {
+            gasLimit: 1000000,
+          })
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, true, 1);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, true, 1);
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -486,22 +486,24 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          roles.manager.address
+          airnodeAddress
         )
       ).to.equal(true);
       // Revoke indefinite whitelisting
       await expect(
-        requesterAuthorizerWithManager
-          .connect(roles.manager)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
+        requesterAuthorizerWithAirnode
+          .connect(airnodeWallet)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false, {
+            gasLimit: 1000000,
+          })
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, false, 0);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, false, 0);
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -509,22 +511,24 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          roles.manager.address
+          airnodeAddress
         )
       ).to.equal(false);
       // Revoking indefinite whitelisting again should have no effect
       await expect(
-        requesterAuthorizerWithManager
-          .connect(roles.manager)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
+        requesterAuthorizerWithAirnode
+          .connect(airnodeWallet)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false, {
+            gasLimit: 1000000,
+          })
       )
-        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, false, 0);
-      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, false, 0);
+      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -532,29 +536,29 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          roles.manager.address
+          airnodeAddress
         )
       ).to.equal(false);
     });
   });
-  context('Sender does not have the indefinite whitelister role and is not the manager address', function () {
+  context('Sender does not have the indefinite whitelister role and is not the Airnode address', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationExtender)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       ).to.be.revertedWith('Not indefinite whitelister');
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       ).to.be.revertedWith('Not indefinite whitelister');
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.randomPerson)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       ).to.be.revertedWith('Not indefinite whitelister');
@@ -564,19 +568,19 @@ describe('setIndefiniteWhitelistStatus', function () {
 
 describe('revokeIndefiniteWhitelistStatus', function () {
   context('setter does not have the indefinite whitelister role', function () {
-    context('setter is not the manager address', function () {
+    context('setter is not the Airnode address', function () {
       it('revokes indefinite whitelist status', async function () {
         // Grant indefinite whitelist status
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         // Revoke the indefinite whitelister role
         await accessControlRegistry
-          .connect(roles.manager)
-          .revokeRole(indefiniteWhitelisterRole, roles.indefiniteWhitelister.address);
+          .connect(airnodeWallet)
+          .revokeRole(indefiniteWhitelisterRole, roles.indefiniteWhitelister.address, { gasLimit: 1000000 });
         // Revoke the indefinite whitelist status
         await expect(
-          requesterAuthorizerWithManager
+          requesterAuthorizerWithAirnode
             .connect(roles.randomPerson)
             .revokeIndefiniteWhitelistStatus(
               airnodeAddress,
@@ -585,7 +589,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
               roles.indefiniteWhitelister.address
             )
         )
-          .to.emit(requesterAuthorizerWithManager, 'RevokedIndefiniteWhitelistStatus')
+          .to.emit(requesterAuthorizerWithAirnode, 'RevokedIndefiniteWhitelistStatus')
           .withArgs(
             airnodeAddress,
             endpointId,
@@ -594,7 +598,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
             roles.randomPerson.address,
             0
           );
-        const whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        const whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -603,7 +607,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
         expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
         // Revoking twice should not emit an event
         await expect(
-          requesterAuthorizerWithManager
+          requesterAuthorizerWithAirnode
             .connect(roles.randomPerson)
             .revokeIndefiniteWhitelistStatus(
               airnodeAddress,
@@ -611,18 +615,18 @@ describe('revokeIndefiniteWhitelistStatus', function () {
               roles.requester.address,
               roles.indefiniteWhitelister.address
             )
-        ).to.not.emit(requesterAuthorizerWithManager, 'RevokedIndefiniteWhitelistStatus');
+        ).to.not.emit(requesterAuthorizerWithAirnode, 'RevokedIndefiniteWhitelistStatus');
       });
     });
-    context('setter is the manager address', function () {
+    context('setter is the Airnode address', function () {
       it('reverts', async function () {
         await accessControlRegistry
-          .connect(roles.manager)
-          .renounceRole(indefiniteWhitelisterRole, roles.manager.address);
+          .connect(airnodeWallet)
+          .renounceRole(indefiniteWhitelisterRole, airnodeAddress, { gasLimit: 1000000 });
         await expect(
-          requesterAuthorizerWithManager
+          requesterAuthorizerWithAirnode
             .connect(roles.randomPerson)
-            .revokeIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, roles.manager.address)
+            .revokeIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, airnodeAddress)
         ).to.be.revertedWith('setter is indefinite whitelister');
       });
     });
@@ -630,7 +634,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
   context('setter has the indefinite whitelister role', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithManager
+        requesterAuthorizerWithAirnode
           .connect(roles.randomPerson)
           .revokeIndefiniteWhitelistStatus(
             airnodeAddress,
@@ -643,18 +647,48 @@ describe('revokeIndefiniteWhitelistStatus', function () {
   });
 });
 
+describe('deriveAdminRole', function () {
+  it('derives admin role for the Airnode', async function () {
+    expect(await requesterAuthorizerWithAirnode.deriveAdminRole(airnodeAddress)).to.equal(adminRole);
+  });
+});
+
+describe('deriveWhitelistExpirationExtenderRole', function () {
+  it('derives whitelist expiration extender role for the Airnode', async function () {
+    expect(await requesterAuthorizerWithAirnode.deriveWhitelistExpirationExtenderRole(airnodeAddress)).to.equal(
+      whitelistExpirationExtenderRole
+    );
+  });
+});
+
+describe('deriveWhitelistExpirationSetterRole', function () {
+  it('derives whitelist expiration setter role for the Airnode', async function () {
+    expect(await requesterAuthorizerWithAirnode.deriveWhitelistExpirationSetterRole(airnodeAddress)).to.equal(
+      whitelistExpirationSetterRole
+    );
+  });
+});
+
+describe('deriveIndefiniteWhitelisterRole', function () {
+  it('derives indefinite whitelister role for the Airnode', async function () {
+    expect(await requesterAuthorizerWithAirnode.deriveIndefiniteWhitelisterRole(airnodeAddress)).to.equal(
+      indefiniteWhitelisterRole
+    );
+  });
+});
+
 describe('requesterIsWhitelisted', function () {
   context('Requester is whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -664,42 +698,42 @@ describe('requesterIsWhitelisted', function () {
     });
     context('Requester is not whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
           )
         ).to.equal(true);
         // Also test the case with two indefinite whitelisters
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.anotherIndefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
           )
         ).to.equal(true);
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false);
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
           )
         ).to.equal(true);
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.anotherIndefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false);
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -711,11 +745,11 @@ describe('requesterIsWhitelisted', function () {
   context('Requester is not whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -726,7 +760,7 @@ describe('requesterIsWhitelisted', function () {
     context('Requester is not whitelisted temporarily', function () {
       it('returns false', async function () {
         expect(
-          await requesterAuthorizerWithManager.requesterIsWhitelisted(
+          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -741,14 +775,14 @@ describe('isAuthorized', function () {
   context('Requester is whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithManager.isAuthorized(
+          await requesterAuthorizerWithAirnode.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
@@ -760,11 +794,11 @@ describe('isAuthorized', function () {
     });
     context('Requester is not whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         expect(
-          await requesterAuthorizerWithManager.isAuthorized(
+          await requesterAuthorizerWithAirnode.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
@@ -778,11 +812,11 @@ describe('isAuthorized', function () {
   context('Requester is not whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithManager
+        await requesterAuthorizerWithAirnode
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithManager.isAuthorized(
+          await requesterAuthorizerWithAirnode.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
@@ -795,7 +829,7 @@ describe('isAuthorized', function () {
     context('Requester is not whitelisted temporarily', function () {
       it('returns false', async function () {
         expect(
-          await requesterAuthorizerWithManager.isAuthorized(
+          await requesterAuthorizerWithAirnode.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
