+++ packages/airnode-protocol/test/authorizers/RequesterAuthorizerWithManager.sol.js
@@ -1,125 +1,135 @@
 /* globals context */
 const hre = require('hardhat');
 const { expect } = require('chai');
-const utils = require('../../utils');
+const utils = require('../utils');
 
 let roles;
-let accessControlRegistry, requesterAuthorizerWithAirnode;
-let requesterAuthorizerWithAirnodeAdminRoleDescription = 'RequesterAuthorizerWithAirnode admin';
+let accessControlRegistry, requesterAuthorizerWithManager;
+let requesterAuthorizerWithManagerAdminRoleDescription = 'RequesterAuthorizerWithManager admin';
 let adminRole, whitelistExpirationExtenderRole, whitelistExpirationSetterRole, indefiniteWhitelisterRole;
-let airnodeAddress, airnodeMnemonic, airnodeWallet;
+let airnodeAddress = utils.generateRandomAddress();
 let endpointId = utils.generateRandomBytes32();
 
 beforeEach(async () => {
   const accounts = await hre.ethers.getSigners();
   roles = {
     deployer: accounts[0],
-    whitelistExpirationExtender: accounts[1],
-    whitelistExpirationSetter: accounts[2],
-    indefiniteWhitelister: accounts[3],
-    anotherIndefiniteWhitelister: accounts[4],
+    manager: accounts[1],
+    whitelistExpirationExtender: accounts[2],
+    whitelistExpirationSetter: accounts[3],
+    indefiniteWhitelister: accounts[4],
+    anotherIndefiniteWhitelister: accounts[5],
     requester: accounts[5],
     randomPerson: accounts[9],
   };
   const accessControlRegistryFactory = await hre.ethers.getContractFactory('AccessControlRegistry', roles.deployer);
   accessControlRegistry = await accessControlRegistryFactory.deploy();
-  const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
-    'RequesterAuthorizerWithAirnode',
+  const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
+    'RequesterAuthorizerWithManager',
     roles.deployer
   );
-  requesterAuthorizerWithAirnode = await requesterAuthorizerWithAirnodeFactory.deploy(
+  requesterAuthorizerWithManager = await requesterAuthorizerWithManagerFactory.deploy(
     accessControlRegistry.address,
-    requesterAuthorizerWithAirnodeAdminRoleDescription
+    requesterAuthorizerWithManagerAdminRoleDescription,
+    roles.manager.address
   );
-  ({ airnodeAddress: airnodeAddress, airnodeMnemonic: airnodeMnemonic } = utils.generateRandomAirnodeWallet());
-  await roles.deployer.sendTransaction({
-    to: airnodeAddress,
-    value: hre.ethers.utils.parseEther('1'),
-  });
-  airnodeWallet = hre.ethers.Wallet.fromMnemonic(airnodeMnemonic).connect(hre.ethers.provider);
-  const airnodeRootRole = await accessControlRegistry.deriveRootRole(airnodeAddress);
-  adminRole = await requesterAuthorizerWithAirnode.deriveAdminRole(airnodeAddress);
-  whitelistExpirationExtenderRole = await requesterAuthorizerWithAirnode.deriveWhitelistExpirationExtenderRole(
-    airnodeAddress
-  );
-  whitelistExpirationSetterRole = await requesterAuthorizerWithAirnode.deriveWhitelistExpirationSetterRole(
-    airnodeAddress
-  );
-  indefiniteWhitelisterRole = await requesterAuthorizerWithAirnode.deriveIndefiniteWhitelisterRole(airnodeAddress);
-  await accessControlRegistry.connect(airnodeWallet).initializeAndGrantRoles(
-    [airnodeRootRole, adminRole, adminRole, adminRole, adminRole],
+  const managerRootRole = await accessControlRegistry.deriveRootRole(roles.manager.address);
+  adminRole = await requesterAuthorizerWithManager.adminRole();
+  whitelistExpirationExtenderRole = await requesterAuthorizerWithManager.whitelistExpirationExtenderRole();
+  whitelistExpirationSetterRole = await requesterAuthorizerWithManager.whitelistExpirationSetterRole();
+  indefiniteWhitelisterRole = await requesterAuthorizerWithManager.indefiniteWhitelisterRole();
+  await accessControlRegistry.connect(roles.manager).initializeAndGrantRoles(
+    [managerRootRole, adminRole, adminRole, adminRole, adminRole],
     [
-      requesterAuthorizerWithAirnodeAdminRoleDescription,
-      await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
-      await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
-      await requesterAuthorizerWithAirnode.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
-      await requesterAuthorizerWithAirnode.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
+      requesterAuthorizerWithManagerAdminRoleDescription,
+      await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
+      await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
+      await requesterAuthorizerWithManager.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
+      await requesterAuthorizerWithManager.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
     ],
     [
-      airnodeAddress, // which will already have been granted the role
+      roles.manager.address, // which will already have been granted the role
       roles.whitelistExpirationExtender.address,
       roles.whitelistExpirationSetter.address,
       roles.indefiniteWhitelister.address,
       roles.anotherIndefiniteWhitelister.address,
-    ],
-    { gasLimit: 1000000 }
+    ]
   );
   // Grant `roles.randomPerson` some invalid roles
   await accessControlRegistry
-    .connect(airnodeWallet)
+    .connect(roles.manager)
     .initializeAndGrantRoles(
-      [airnodeRootRole, airnodeRootRole, airnodeRootRole, airnodeRootRole],
+      [managerRootRole, managerRootRole, managerRootRole, managerRootRole],
       [
         Math.random(),
-        await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
-        await requesterAuthorizerWithAirnode.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
-        await requesterAuthorizerWithAirnode.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
+        await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_EXTENDER_ROLE_DESCRIPTION(),
+        await requesterAuthorizerWithManager.WHITELIST_EXPIRATION_SETTER_ROLE_DESCRIPTION(),
+        await requesterAuthorizerWithManager.INDEFINITE_WHITELISTER_ROLE_DESCRIPTION(),
       ],
-      [roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address],
-      { gasLimit: 1000000 }
+      [roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address, roles.randomPerson.address]
     );
 });
 
 describe('constructor', function () {
   context('AccessControlRegistry address is not zero', function () {
     context('Admin role description string is not empty', function () {
-      it('constructs', async function () {
-        const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
-          'RequesterAuthorizerWithAirnode',
-          roles.deployer
-        );
-        requesterAuthorizerWithAirnode = await requesterAuthorizerWithAirnodeFactory.deploy(
-          accessControlRegistry.address,
-          requesterAuthorizerWithAirnodeAdminRoleDescription
-        );
-        expect(await requesterAuthorizerWithAirnode.accessControlRegistry()).to.equal(accessControlRegistry.address);
-        expect(await requesterAuthorizerWithAirnode.adminRoleDescription()).to.equal(
-          requesterAuthorizerWithAirnodeAdminRoleDescription
-        );
+      context('Manager address is not zero', function () {
+        it('constructs', async function () {
+          const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
+            'RequesterAuthorizerWithManager',
+            roles.deployer
+          );
+          requesterAuthorizerWithManager = await requesterAuthorizerWithManagerFactory.deploy(
+            accessControlRegistry.address,
+            requesterAuthorizerWithManagerAdminRoleDescription,
+            roles.manager.address
+          );
+          expect(await requesterAuthorizerWithManager.accessControlRegistry()).to.equal(accessControlRegistry.address);
+          expect(await requesterAuthorizerWithManager.adminRoleDescription()).to.equal(
+            requesterAuthorizerWithManagerAdminRoleDescription
+          );
+          expect(await requesterAuthorizerWithManager.manager()).to.equal(roles.manager.address);
+        });
+      });
+      context('Manager address is zero', function () {
+        it('reverts', async function () {
+          const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
+            'RequesterAuthorizerWithManager',
+            roles.deployer
+          );
+          await expect(
+            requesterAuthorizerWithManagerFactory.deploy(
+              accessControlRegistry.address,
+              requesterAuthorizerWithManagerAdminRoleDescription,
+              hre.ethers.constants.AddressZero
+            )
+          ).to.be.revertedWith('Manager address zero');
+        });
       });
     });
     context('Admin role description string is empty', function () {
       it('reverts', async function () {
-        const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
-          'RequesterAuthorizerWithAirnode',
+        const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
+          'RequesterAuthorizerWithManager',
           roles.deployer
         );
         await expect(
-          requesterAuthorizerWithAirnodeFactory.deploy(accessControlRegistry.address, '')
+          requesterAuthorizerWithManagerFactory.deploy(accessControlRegistry.address, '', roles.manager.address)
         ).to.be.revertedWith('Admin role description empty');
       });
     });
   });
   context('AccessControlRegistry address is zero', function () {
     it('reverts', async function () {
-      const requesterAuthorizerWithAirnodeFactory = await hre.ethers.getContractFactory(
-        'RequesterAuthorizerWithAirnode',
+      const requesterAuthorizerWithManagerFactory = await hre.ethers.getContractFactory(
+        'RequesterAuthorizerWithManager',
         roles.deployer
       );
       await expect(
-        requesterAuthorizerWithAirnodeFactory.deploy(
+        requesterAuthorizerWithManagerFactory.deploy(
           hre.ethers.constants.AddressZero,
-          requesterAuthorizerWithAirnodeAdminRoleDescription
+          requesterAuthorizerWithManagerAdminRoleDescription,
+          roles.manager.address
         )
       ).to.be.revertedWith('ACR address zero');
     });
@@ -131,7 +141,7 @@ describe('extendWhitelistExpiration', function () {
     context('Timestamp extends whitelist expiration', function () {
       it('extends whitelist expiration', async function () {
         let whitelistStatus;
-        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -140,11 +150,11 @@ describe('extendWhitelistExpiration', function () {
         expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
         const expirationTimestamp = 1000;
         await expect(
-          requesterAuthorizerWithAirnode
+          requesterAuthorizerWithManager
             .connect(roles.whitelistExpirationExtender)
             .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
         )
-          .to.emit(requesterAuthorizerWithAirnode, 'ExtendedWhitelistExpiration')
+          .to.emit(requesterAuthorizerWithManager, 'ExtendedWhitelistExpiration')
           .withArgs(
             airnodeAddress,
             endpointId,
@@ -152,7 +162,7 @@ describe('extendWhitelistExpiration', function () {
             roles.whitelistExpirationExtender.address,
             expirationTimestamp
           );
-        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -164,21 +174,21 @@ describe('extendWhitelistExpiration', function () {
     context('Timestamp does not extend whitelist expiration', function () {
       it('reverts', async function () {
         await expect(
-          requesterAuthorizerWithAirnode
+          requesterAuthorizerWithManager
             .connect(roles.whitelistExpirationExtender)
             .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
         ).to.be.revertedWith('Does not extend expiration');
       });
     });
   });
-  context('Sender is the Airnode address', function () {
+  context('Sender is the manager address', function () {
     context('Timestamp extends whitelist expiration', function () {
       it('extends whitelist expiration', async function () {
         await accessControlRegistry
-          .connect(airnodeWallet)
-          .renounceRole(whitelistExpirationExtenderRole, airnodeAddress, { gasLimit: 1000000 });
+          .connect(roles.manager)
+          .renounceRole(whitelistExpirationExtenderRole, roles.manager.address);
         let whitelistStatus;
-        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -187,15 +197,13 @@ describe('extendWhitelistExpiration', function () {
         expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
         const expirationTimestamp = 1000;
         await expect(
-          requesterAuthorizerWithAirnode
-            .connect(airnodeWallet)
-            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp, {
-              gasLimit: 1000000,
-            })
+          requesterAuthorizerWithManager
+            .connect(roles.manager)
+            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
         )
-          .to.emit(requesterAuthorizerWithAirnode, 'ExtendedWhitelistExpiration')
-          .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, expirationTimestamp);
-        whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+          .to.emit(requesterAuthorizerWithManager, 'ExtendedWhitelistExpiration')
+          .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, expirationTimestamp);
+        whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -207,30 +215,30 @@ describe('extendWhitelistExpiration', function () {
     context('Timestamp does not extend whitelist expiration', function () {
       it('reverts', async function () {
         await accessControlRegistry
-          .connect(airnodeWallet)
-          .renounceRole(whitelistExpirationExtenderRole, airnodeAddress, { gasLimit: 1000000 });
+          .connect(roles.manager)
+          .renounceRole(whitelistExpirationExtenderRole, roles.manager.address);
         await expect(
-          requesterAuthorizerWithAirnode
-            .connect(airnodeWallet)
-            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0, { gasLimit: 1000000 })
+          requesterAuthorizerWithManager
+            .connect(roles.manager)
+            .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
         ).to.be.revertedWith('Does not extend expiration');
       });
     });
   });
-  context('Sender does not have the whitelist extender role and is not the Airnode address', function () {
+  context('Sender does not have the whitelist extender role and is not the manager address', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 1000)
       ).to.be.revertedWith('Not expiration extender');
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 1000)
       ).to.be.revertedWith('Not expiration extender');
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.randomPerson)
           .extendWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 1000)
       ).to.be.revertedWith('Not expiration extender');
@@ -244,11 +252,11 @@ describe('setWhitelistExpiration', function () {
       let whitelistStatus;
       const expirationTimestamp = 1000;
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
+        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
         .withArgs(
           airnodeAddress,
           endpointId,
@@ -256,7 +264,7 @@ describe('setWhitelistExpiration', function () {
           roles.whitelistExpirationSetter.address,
           expirationTimestamp
         );
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -264,13 +272,13 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(expirationTimestamp);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
+        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.whitelistExpirationSetter.address, 0);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -279,23 +287,21 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
     });
   });
-  context('Sender is the Airnode address', function () {
+  context('Sender is tha manager address', function () {
     it('sets whitelist expiration', async function () {
       await accessControlRegistry
-        .connect(airnodeWallet)
-        .renounceRole(whitelistExpirationSetterRole, airnodeAddress, { gasLimit: 1000000 });
+        .connect(roles.manager)
+        .renounceRole(whitelistExpirationSetterRole, roles.manager.address);
       let whitelistStatus;
       const expirationTimestamp = 1000;
       await expect(
-        requesterAuthorizerWithAirnode
-          .connect(airnodeWallet)
-          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp, {
-            gasLimit: 1000000,
-          })
+        requesterAuthorizerWithManager
+          .connect(roles.manager)
+          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, expirationTimestamp)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, expirationTimestamp);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, expirationTimestamp);
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -303,13 +309,13 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(expirationTimestamp);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       await expect(
-        requesterAuthorizerWithAirnode
-          .connect(airnodeWallet)
-          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0, { gasLimit: 1000000 })
+        requesterAuthorizerWithManager
+          .connect(roles.manager)
+          .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetWhitelistExpiration')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, 0);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithManager, 'SetWhitelistExpiration')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, 0);
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -318,20 +324,20 @@ describe('setWhitelistExpiration', function () {
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
     });
   });
-  context('Sender does not have the whitelist expiration setter role and is not the Airnode address', function () {
+  context('Sender does not have the whitelist expiration setter role and is not the manager address', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationExtender)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       ).to.be.revertedWith('Not expiration setter');
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       ).to.be.revertedWith('Not expiration setter');
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.randomPerson)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 0)
       ).to.be.revertedWith('Not expiration setter');
@@ -345,13 +351,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       let whitelistStatus;
       // Whitelist indefinitely
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, true, 1);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -359,7 +365,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -368,13 +374,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(true);
       // Whitelisting indefinitely again should have no effect
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, true, 1);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -382,7 +388,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -391,13 +397,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(true);
       // Revoke indefinite whitelisting
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, false, 0);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -405,7 +411,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -414,13 +420,13 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(false);
       // Revoking indefinite whitelisting again should have no effect
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
         .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.indefiniteWhitelister.address, false, 0);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -428,7 +434,7 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
@@ -437,23 +443,19 @@ describe('setIndefiniteWhitelistStatus', function () {
       ).to.equal(false);
     });
   });
-  context('Sender is the Airnode address', function () {
+  context('Sender is the manager address', function () {
     it('sets indefinite whitelist status', async function () {
-      await accessControlRegistry
-        .connect(airnodeWallet)
-        .renounceRole(indefiniteWhitelisterRole, airnodeAddress, { gasLimit: 1000000 });
+      await accessControlRegistry.connect(roles.manager).renounceRole(indefiniteWhitelisterRole, roles.manager.address);
       let whitelistStatus;
       // Whitelist indefinitely
       await expect(
-        requesterAuthorizerWithAirnode
-          .connect(airnodeWallet)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true, {
-            gasLimit: 1000000,
-          })
+        requesterAuthorizerWithManager
+          .connect(roles.manager)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, true, 1);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, true, 1);
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -461,24 +463,22 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          airnodeAddress
+          roles.manager.address
         )
       ).to.equal(true);
       // Whitelisting indefinitely again should have no effect
       await expect(
-        requesterAuthorizerWithAirnode
-          .connect(airnodeWallet)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true, {
-            gasLimit: 1000000,
-          })
+        requesterAuthorizerWithManager
+          .connect(roles.manager)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, true, 1);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, true, 1);
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -486,24 +486,22 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(1);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          airnodeAddress
+          roles.manager.address
         )
       ).to.equal(true);
       // Revoke indefinite whitelisting
       await expect(
-        requesterAuthorizerWithAirnode
-          .connect(airnodeWallet)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false, {
-            gasLimit: 1000000,
-          })
+        requesterAuthorizerWithManager
+          .connect(roles.manager)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, false, 0);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, false, 0);
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -511,24 +509,22 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          airnodeAddress
+          roles.manager.address
         )
       ).to.equal(false);
       // Revoking indefinite whitelisting again should have no effect
       await expect(
-        requesterAuthorizerWithAirnode
-          .connect(airnodeWallet)
-          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false, {
-            gasLimit: 1000000,
-          })
+        requesterAuthorizerWithManager
+          .connect(roles.manager)
+          .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false)
       )
-        .to.emit(requesterAuthorizerWithAirnode, 'SetIndefiniteWhitelistStatus')
-        .withArgs(airnodeAddress, endpointId, roles.requester.address, airnodeAddress, false, 0);
-      whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        .to.emit(requesterAuthorizerWithManager, 'SetIndefiniteWhitelistStatus')
+        .withArgs(airnodeAddress, endpointId, roles.requester.address, roles.manager.address, false, 0);
+      whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
         airnodeAddress,
         endpointId,
         roles.requester.address
@@ -536,29 +532,29 @@ describe('setIndefiniteWhitelistStatus', function () {
       expect(whitelistStatus.expirationTimestamp).to.equal(0);
       expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
       expect(
-        await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
+        await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToSetterToIndefiniteWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address,
-          airnodeAddress
+          roles.manager.address
         )
       ).to.equal(false);
     });
   });
-  context('Sender does not have the indefinite whitelister role and is not the Airnode address', function () {
+  context('Sender does not have the indefinite whitelister role and is not the manager address', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationExtender)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       ).to.be.revertedWith('Not indefinite whitelister');
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       ).to.be.revertedWith('Not indefinite whitelister');
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.randomPerson)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true)
       ).to.be.revertedWith('Not indefinite whitelister');
@@ -568,19 +564,19 @@ describe('setIndefiniteWhitelistStatus', function () {
 
 describe('revokeIndefiniteWhitelistStatus', function () {
   context('setter does not have the indefinite whitelister role', function () {
-    context('setter is not the Airnode address', function () {
+    context('setter is not the manager address', function () {
       it('revokes indefinite whitelist status', async function () {
         // Grant indefinite whitelist status
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         // Revoke the indefinite whitelister role
         await accessControlRegistry
-          .connect(airnodeWallet)
-          .revokeRole(indefiniteWhitelisterRole, roles.indefiniteWhitelister.address, { gasLimit: 1000000 });
+          .connect(roles.manager)
+          .revokeRole(indefiniteWhitelisterRole, roles.indefiniteWhitelister.address);
         // Revoke the indefinite whitelist status
         await expect(
-          requesterAuthorizerWithAirnode
+          requesterAuthorizerWithManager
             .connect(roles.randomPerson)
             .revokeIndefiniteWhitelistStatus(
               airnodeAddress,
@@ -589,7 +585,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
               roles.indefiniteWhitelister.address
             )
         )
-          .to.emit(requesterAuthorizerWithAirnode, 'RevokedIndefiniteWhitelistStatus')
+          .to.emit(requesterAuthorizerWithManager, 'RevokedIndefiniteWhitelistStatus')
           .withArgs(
             airnodeAddress,
             endpointId,
@@ -598,7 +594,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
             roles.randomPerson.address,
             0
           );
-        const whitelistStatus = await requesterAuthorizerWithAirnode.airnodeToEndpointIdToRequesterToWhitelistStatus(
+        const whitelistStatus = await requesterAuthorizerWithManager.airnodeToEndpointIdToRequesterToWhitelistStatus(
           airnodeAddress,
           endpointId,
           roles.requester.address
@@ -607,7 +603,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
         expect(whitelistStatus.indefiniteWhitelistCount).to.equal(0);
         // Revoking twice should not emit an event
         await expect(
-          requesterAuthorizerWithAirnode
+          requesterAuthorizerWithManager
             .connect(roles.randomPerson)
             .revokeIndefiniteWhitelistStatus(
               airnodeAddress,
@@ -615,18 +611,18 @@ describe('revokeIndefiniteWhitelistStatus', function () {
               roles.requester.address,
               roles.indefiniteWhitelister.address
             )
-        ).to.not.emit(requesterAuthorizerWithAirnode, 'RevokedIndefiniteWhitelistStatus');
+        ).to.not.emit(requesterAuthorizerWithManager, 'RevokedIndefiniteWhitelistStatus');
       });
     });
-    context('setter is the Airnode address', function () {
+    context('setter is the manager address', function () {
       it('reverts', async function () {
         await accessControlRegistry
-          .connect(airnodeWallet)
-          .renounceRole(indefiniteWhitelisterRole, airnodeAddress, { gasLimit: 1000000 });
+          .connect(roles.manager)
+          .renounceRole(indefiniteWhitelisterRole, roles.manager.address);
         await expect(
-          requesterAuthorizerWithAirnode
+          requesterAuthorizerWithManager
             .connect(roles.randomPerson)
-            .revokeIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, airnodeAddress)
+            .revokeIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, roles.manager.address)
         ).to.be.revertedWith('setter is indefinite whitelister');
       });
     });
@@ -634,7 +630,7 @@ describe('revokeIndefiniteWhitelistStatus', function () {
   context('setter has the indefinite whitelister role', function () {
     it('reverts', async function () {
       await expect(
-        requesterAuthorizerWithAirnode
+        requesterAuthorizerWithManager
           .connect(roles.randomPerson)
           .revokeIndefiniteWhitelistStatus(
             airnodeAddress,
@@ -647,48 +643,18 @@ describe('revokeIndefiniteWhitelistStatus', function () {
   });
 });
 
-describe('deriveAdminRole', function () {
-  it('derives admin role for the Airnode', async function () {
-    expect(await requesterAuthorizerWithAirnode.deriveAdminRole(airnodeAddress)).to.equal(adminRole);
-  });
-});
-
-describe('deriveWhitelistExpirationExtenderRole', function () {
-  it('derives whitelist expiration extender role for the Airnode', async function () {
-    expect(await requesterAuthorizerWithAirnode.deriveWhitelistExpirationExtenderRole(airnodeAddress)).to.equal(
-      whitelistExpirationExtenderRole
-    );
-  });
-});
-
-describe('deriveWhitelistExpirationSetterRole', function () {
-  it('derives whitelist expiration setter role for the Airnode', async function () {
-    expect(await requesterAuthorizerWithAirnode.deriveWhitelistExpirationSetterRole(airnodeAddress)).to.equal(
-      whitelistExpirationSetterRole
-    );
-  });
-});
-
-describe('deriveIndefiniteWhitelisterRole', function () {
-  it('derives indefinite whitelister role for the Airnode', async function () {
-    expect(await requesterAuthorizerWithAirnode.deriveIndefiniteWhitelisterRole(airnodeAddress)).to.equal(
-      indefiniteWhitelisterRole
-    );
-  });
-});
-
 describe('requesterIsWhitelisted', function () {
   context('Requester is whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -698,42 +664,42 @@ describe('requesterIsWhitelisted', function () {
     });
     context('Requester is not whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
           )
         ).to.equal(true);
         // Also test the case with two indefinite whitelisters
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.anotherIndefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
           )
         ).to.equal(true);
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false);
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
           )
         ).to.equal(true);
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.anotherIndefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, false);
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -745,11 +711,11 @@ describe('requesterIsWhitelisted', function () {
   context('Requester is not whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -760,7 +726,7 @@ describe('requesterIsWhitelisted', function () {
     context('Requester is not whitelisted temporarily', function () {
       it('returns false', async function () {
         expect(
-          await requesterAuthorizerWithAirnode.requesterIsWhitelisted(
+          await requesterAuthorizerWithManager.requesterIsWhitelisted(
             airnodeAddress,
             endpointId,
             roles.requester.address
@@ -775,14 +741,14 @@ describe('isAuthorized', function () {
   context('Requester is whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithAirnode.isAuthorized(
+          await requesterAuthorizerWithManager.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
@@ -794,11 +760,11 @@ describe('isAuthorized', function () {
     });
     context('Requester is not whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.indefiniteWhitelister)
           .setIndefiniteWhitelistStatus(airnodeAddress, endpointId, roles.requester.address, true);
         expect(
-          await requesterAuthorizerWithAirnode.isAuthorized(
+          await requesterAuthorizerWithManager.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
@@ -812,11 +778,11 @@ describe('isAuthorized', function () {
   context('Requester is not whitelisted indefinitely', function () {
     context('Requester is whitelisted temporarily', function () {
       it('returns true', async function () {
-        await requesterAuthorizerWithAirnode
+        await requesterAuthorizerWithManager
           .connect(roles.whitelistExpirationSetter)
           .setWhitelistExpiration(airnodeAddress, endpointId, roles.requester.address, 2000000000);
         expect(
-          await requesterAuthorizerWithAirnode.isAuthorized(
+          await requesterAuthorizerWithManager.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
@@ -829,7 +795,7 @@ describe('isAuthorized', function () {
     context('Requester is not whitelisted temporarily', function () {
       it('returns false', async function () {
         expect(
-          await requesterAuthorizerWithAirnode.isAuthorized(
+          await requesterAuthorizerWithManager.isAuthorized(
             utils.generateRandomBytes32(),
             airnodeAddress,
             endpointId,
